Telemetry cleanup
