+++ Cargo.toml
@@ -18,19 +18,16 @@ sqlx = { version = "0.4.2", default-features = false, features = [
     "json",
 ] }
 
-
+actix-cors = "0.5.4"
 config = "0.10.1"
-chrono = "0.4.19"
+chrono = { version = "0.4.19", features = [ "serde" ] }
 serde = { version = "1.0.120", features = [ "derive" ] }
 uuid = { version = "0.8.2", features = [ "v4", "serde" ] }
 
-tracing = "0.1.22"
-tracing-futures = "0.2.4"
-tracing-subscriber = { version = "0.2.12", features = ["registry", "env-filter"] }
-tracing-bunyan-formatter = "0.1.7"
-tracing-log = "0.1.1"
-tracing-actix-web = "0.2.1"
-actix-cors = "0.5.4"
+tracing = "0.1"
+tracing-actix-web = "0.2"
+tracing-futures = "0.2"
+tracing-subscriber = { version = "0.2", features = ["registry", "env-filter", "tracing-log", "json"] }
 
 [dev-dependencies]
 lazy_static = "1.4.0"
